+++ benchmarks/datascript-bench/bench.clj
@@ -153,8 +153,8 @@
 
 (def default-versions
   [
-   ;; ["latest" "datomic"]
-   ;; ["latest" "datascript"]
+   ;["latest" "datomic"]
+   ;["latest" "datascript"]
    ["0.9.14" "datalevin"]
    ["latest" "datalevin"]
    ])